Refactor Android platform export structure.
